[RED] Swap first two elements when needed (split flow+assignment)
To make the new test pass we swap the first two elements of the list if
the second is bigger than the first. However this will make the empty
and one-element list testcases fail with IndexErrors, that has to be
fixed. The added if conditional is the split flow transformation and
modifying elements of the list is assignment.